Moved formRowView class to ui components, also bookmarks were updated with this change.
