+++ packages/ckeditor5-bookmark/tests/ui/bookmarkformview.js
@@ -12,6 +12,7 @@ import KeystrokeHandler from '@ckeditor/ckeditor5-utils/src/keystrokehandler.js'
 import FocusTracker from '@ckeditor/ckeditor5-utils/src/focustracker.js';
 import FocusCycler from '@ckeditor/ckeditor5-ui/src/focuscycler.js';
 import FormHeaderView from '@ckeditor/ckeditor5-ui/src/formheader/formheaderview.js';
+import FormRowView from '@ckeditor/ckeditor5-ui/src/formrow/formrowview.js';
 import ViewCollection from '@ckeditor/ckeditor5-ui/src/viewcollection.js';
 import testUtils from '@ckeditor/ckeditor5-core/tests/_utils/utils.js';
 
@@ -53,12 +54,13 @@ describe( 'BookmarkFormView', () => {
 
 			const formRowView = view.children.get( 1 );
 
+			expect( formRowView ).to.be.instanceOf( FormRowView );
 			expect( formRowView.element.classList.contains( 'ck' ) ).to.true;
 			expect( formRowView.element.classList.contains( 'ck-form__row' ) ).to.true;
 			expect( formRowView.element.classList.contains( 'ck-form__row_with-submit' ) ).to.true;
 			expect( formRowView.element.classList.contains( 'ck-form__row_large-top-padding' ) ).to.true;
-			expect( formRowView.template.children[ 0 ] ).to.equal( view.idInputView );
-			expect( formRowView.template.children[ 1 ] ).to.equal( view.saveButtonView );
+			expect( formRowView.children.get( 0 ) ).to.equal( view.idInputView );
+			expect( formRowView.children.get( 1 ) ).to.equal( view.saveButtonView );
 
 			const formHeaderView = view.children.get( 0 );
